Reorganized directory structure to a much cleaner arrangement; everything still works
