+++ src/privateer/analysis.py
@@ -1,5 +1,5 @@
 import xml.etree.ElementTree as etree
-from privateer.privateer_core import get_annotated_glycans,get_annotated_glycans_hierarchical
+from privateer.pyprivateer import get_annotated_glycans,get_annotated_glycans_hierarchical
 
 def print_glycosidic_torsions ( pdb_filename = "", first="ASN", second="NAG" ) :
