+++ src/privateer/cpp/privateer-blobs.cpp
