+++ src/privateer/cpp/privateer-blobs.h
