+++ src/privateer/cpp/privateer-composition.h
