+++ src/privateer/cpp/privateer-lib.h
